app/requirements.txt
crewai
langchain>=0.2.6,<0.3.0
boto3>=1.34.0
langchain-aws>=0.1.0
yfinance>=0.2.43
streamlit>=1.37.0
pandas>=2.2.2,<2.3.0
plotly
scipy>=1.13.1,<1.14.0
numpy>=1.26.0,<2.0.0
textblob
pandas-ta-classic
requests>=2.31.0
pyjwt>=2.8.0
redis>=4.5.5
python-dotenv>=1.0.0
